commit changepment de type de variable pour la classe Client

+++ src/main/java/com/example/gestioncontrat/bean/Client.java
@@ -10,8 +10,8 @@ public class Client {
     private String nom;
     private String prenom;
     private String adresse;
-    private int NumTel;
-    private int NumCompte;
+    private Long NumTel;
+    private Long NumCompte;
     private String NumCin;
     @OneToMany(mappedBy = "client")
     private List<LigneClient> ligneClients;
@@ -56,19 +56,19 @@ public class Client {
         this.adresse = adresse;
     }
 
-    public int getNumTel() {
+    public Long getNumTel() {
         return NumTel;
     }
 
-    public void setNumTel(int numTel) {
+    public void setNumTel(Long numTel) {
         NumTel = numTel;
     }
 
-    public int getNumCompte() {
+    public Long getNumCompte() {
         return NumCompte;
     }
 
-    public void setNumCompte(int numCompte) {
+    public void setNumCompte(Long numCompte) {
         NumCompte = numCompte;
     }
